# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quookly API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/accounts/bootstrap": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Bootstrap State",
        "description": "Whether this instance still needs its first admin. Public by necessity.",
        "operationId": "getBootstrapState",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapState"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Bootstrap Admin",
        "description": "Claim a fresh instance by creating its first, admin, account.",
        "operationId": "bootstrapAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Authenticated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Register Cook",
        "description": "Create an account.",
        "operationId": "registerCook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Authenticated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/accounts/sign-in": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Sign In",
        "description": "Exchange credentials for a token.",
        "operationId": "signIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Authenticated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/setup/declarations/{step}": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Declare Step",
        "description": "Answer a setup question outright \u2014 \"nobody has any\", \"the defaults are fine\".\n\nThe one stored part of setup. Without it a household where nobody has a dietary\nrestriction is indistinguishable from one nobody has been asked about, and setup\nwould go on asking forever (FR-15).",
        "operationId": "declareStep",
        "parameters": [
          {
            "name": "step",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SetupStep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupProgressView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/locale": {
      "put": {
        "tags": [
          "setup"
        ],
        "summary": "Choose Locale",
        "description": "Remember the language this cook reads in, wherever they sign in from.",
        "operationId": "chooseLocale",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocaleChoice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupProgressView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences/units/{kind}": {
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Choose Unit",
        "description": "Choose the unit for one kind of ingredient.",
        "operationId": "chooseUnit",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IngredientKind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitChoice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitPreferenceView"
                  },
                  "title": "Response Chooseunit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Authenticated": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "cook": {
            "$ref": "#/components/schemas/Cook"
          }
        },
        "type": "object",
        "required": [
          "token",
          "cook"
        ],
        "title": "Authenticated",
        "description": "A signed-in cook and the token proving it."
      },
      "BootstrapState": {
        "properties": {
          "required": {
            "type": "boolean",
            "title": "Required"
          }
        },
        "type": "object",
        "required": [
          "required"
        ],
        "title": "BootstrapState",
        "description": "Whether the instance still needs its first admin."
      },
...
      "Cook": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "is_admin",
          "registered_at"
        ],
        "title": "Cook",
        "description": "A cook account, safe to pass anywhere in the application.\n\nThe password hash is deliberately absent. Anything needing it asks for a\n`StoredCredential`, which keeps credentials on the authentication path instead of\nriding along in every value that happens to describe a cook."
      },
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials",
        "description": "A request to sign in. No length rule \u2014 an old password is still the password."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngredientKind": {
        "type": "string",
        "enum": [
          "liquid",
          "powder",
          "solid",
          "countable"
        ],
        "title": "IngredientKind",
        "description": "What sort of thing this is, for the purpose of measuring it.\n\nThis is the axis a cook's unit preferences run along: powders in grams, liquids in\nmillilitres (UC-6.2). It is deliberately coarse \u2014 it exists to choose a unit, not to\nclassify food."
      },
...
      "LocaleChoice": {
        "properties": {
          "locale": {
            "type": "string",
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "locale"
        ],
        "title": "LocaleChoice",
        "description": "The language a cook reads in."
      },
...
      "Registration": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 12,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "display_name",
          "password"
        ],
        "title": "Registration",
        "description": "A request to create an account.\n\nThe password rule lives here rather than in the manager: it constrains the shape of\nthe input, so it is the contract's business, and stating it here means the API\nrejects a short password before any work is done."
      },
      "SetupProgressView": {
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepStatusView"
            },
            "type": "array",
            "title": "Steps"
          },
          "next_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SetupStep"
              },
              {
                "type": "null"
              }
            ]
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          }
        },
        "type": "object",
        "required": [
          "steps",
          "next_step",
          "complete"
        ],
        "title": "SetupProgressView",
        "description": "What is still missing from the setup, and what to do next (UC-10.3)."
      },
      "SetupStep": {
        "type": "string",
        "enum": [
          "household",
          "constraints",
          "units",
          "locale"
        ],
        "title": "SetupStep",
        "description": "The things worth establishing, in the order they are worth asking for.\n\nWho is eating comes first because it is what the product is about; everything after it\nis a preference about how the answer is shown."
      },
...
      "StepStatusView": {
        "properties": {
          "step": {
            "$ref": "#/components/schemas/SetupStep"
          },
          "done": {
            "type": "boolean",
            "title": "Done"
          },
          "declared": {
            "type": "boolean",
            "title": "Declared"
          }
        },
        "type": "object",
        "required": [
          "step",
          "done",
          "declared"
        ],
        "title": "StepStatusView"
      },
      "UnitChoice": {
        "properties": {
          "unit": {
            "type": "string",
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "unit"
        ],
        "title": "UnitChoice",
        "description": "The unit a cook wants this kind of ingredient shown in."
      },
      "UnitPreferenceView": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/IngredientKind"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "chosen": {
            "type": "boolean",
            "title": "Chosen"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "unit",
          "chosen"
        ],
        "title": "UnitPreferenceView",
        "description": "One kind of ingredient and the unit this cook reads it in.\n\n`chosen` separates a preference from a default. Everybody has defaults, and having\nthem is not an answer \u2014 which is what lets guided setup tell the two apart (FR-15)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}